+++ src/parasail_nomalloc.h
@@ -5,7 +5,7 @@
 
 /* Bytes required for workspace of parasail_sw_striped_profile_avx2_256_16_nomalloc.
    s1Len is query length (profile->s1Len). Internal alignment padding included. */
-size_t parasail_nomalloc_sw_striped_profile_avx2_256_16_workspace_bytes(int s1Len);
+uint parasail_nomalloc_sw_striped_profile_avx2_256_16_workspace_bytes(int s1Len);
 
 /* Score-only striped SW (16-bit) using caller-provided workspace (no malloc/free).
    Returns alignment score (same as parasail_result_t::score from the malloc version).
